requirements.txt
python-fasthtml==0.12.22
anthropic==0.40.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36